Deprecated getCode time argument to be anything but DateTimeInterface (#72)
It was extremely unclear what to do with this argument as well as it was
not documented. Turns out this argument is supposed to be divided by the
codePeriod (30 seconds) before passing along. With a DateTimeInterface
as argument, it's impossible to pass an incorrect value and it will
always ensure that internally this is done right.